security/manager/ssl/tests/mochitest/mixedcontent/test_dynUnsecureBackground.html

<!DOCTYPE HTML>
<html>
<head>
  <title>body.background changes to unsecure test</title>
  <script type="text/javascript" src="/MochiKit/Base.js"></script>
  <script type="text/javascript" src="/MochiKit/DOM.js"></script>
  <script type="text/javascript" src="/MochiKit/Style.js"></script>
  <script type="text/javascript" src="/MochiKit/Signal.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mixedContentTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <script class="testbody" type="text/javascript">

  // This test, as is, equals to https://kuix.de/misc/test17/358438.php

  function runTest()
  {
    isSecurityState("secure");
    document.body.background =
      "http://example.com/tests/security/ssl/mixedcontent/moonsurface.jpg";

    waitForSecurityState("broken", function() 
    {
      isSecurityState("broken", "document.body.background='http://...' changed to broken");
      finish();
    });
  }

  function afterNavigationTest()
  {
    is(document.body.background,
      "https://example.com/tests/security/ssl/mixedcontent/moonsurface.jpg",
      "document backround secure again");
    isSecurityState("secure", "secure after re-navigation");
    finish();
  }

  </script>
</head>

<body background="https://example.com/tests/security/ssl/mixedcontent/moonsurface.jpg">
</body>
</html>
